+++ src/app/pages/estaciones/estaciones.component.ts
@@ -7,6 +7,9 @@ import { ToastrService } from "ngx-toastr";
 import { Observador } from "../../models/observador";
 import { DataTableDirective } from "angular-datatables";
 import { FileUploader } from "ng2-file-upload";
+import { Location } from "@angular/common";
+import { Pais } from "src/app/models/pais";
+import { Division } from "src/app/models/division";
 
 /**
  * Componente para la pagina de edición de estaciones.
@@ -64,16 +67,30 @@ export class EstacionesComponent implements OnInit, OnDestroy {
         pais: ""
     }
 
-    isDtInitialized: boolean = false
 
+    pais = "";
+    division1 = "";
+    division2 = "";
+    division3 = "";
+
+    paises: Pais[] = [];
+    divisiones1: Division[] = [];
+    divisiones2: Division[] = [];
+    divisiones3: Division[] = [];
+
+
+    isDtInitialized: boolean = false
+    public location!: Location;
+    isAdmin = false;
     /**
      * Constructor
      */
     constructor(
+        location: Location,
         private dbService: DbService,
         private tService: ToastrService
     ) {
-
+        this.location = location;
 
         this.uploader = new FileUploader({
             url: dbService.dbURL + "estacion/updatePicture",
@@ -95,6 +112,15 @@ export class EstacionesComponent implements OnInit, OnDestroy {
      * Obtencion de las estaciónes desde la base de datos
      */
     ngOnInit(): void {
+        console.log(this.location.path())
+        let titlee = this.location.prepareExternalUrl(this.location.path());
+        if (titlee.charAt(0) === "#") {
+            titlee = titlee.slice(1);
+        }
+        titlee = titlee.split("/")[1]
+        if (titlee === "admin-layout") {
+            this.isAdmin = true;
+        }
 
         this.isCancel = false;
         this.isError = false;
@@ -154,21 +180,34 @@ export class EstacionesComponent implements OnInit, OnDestroy {
      * @param estacion estacion a editar
      */
     editarEstacion(estacion: any): void {
+        console.log(estacion)
         this.dbService.getObservadores(estacion)
             .subscribe((data: any) => {
                 this.usuarios = (data as any);
                 this.dbService.getFotoEstacion(this.estacion)
                     .subscribe(
                         (data: any) => {
-                            if(data.foto != undefined){
+                            if (data.foto != undefined) {
                                 this.estacion.foto = data.foto
                             }
                         }
                     );
             });
+        this.dbService.getPaises()
+            .subscribe((data: any) => {
+                this.paises = (data as any);
+                console.log(this.paises);
+            });
         this.estacion = estacion;
         this.estacion.latitud = estacion.posicion.coordinates[0];
         this.estacion.longitud = estacion.posicion.coordinates[1];
+        this.getDivisiones(estacion.Division.Pai.id)
+        this.getDivDivisiones(estacion.division1.id, 1)
+        this.getDivDivisiones(estacion.division2.id, 2)
+        this.pais = estacion.Division.Pai.id;
+        this.division1 = estacion.division1.id;
+        this.division2 = estacion.division2.id;
+        this.division3 = estacion.division3.id;
         const table = (<HTMLInputElement>document.getElementById("table"));
         const form = (<HTMLInputElement>document.getElementById("form-estacion"));
         table.style.display = "none";
@@ -181,14 +220,16 @@ export class EstacionesComponent implements OnInit, OnDestroy {
      */
     deleteEstacion(estacion: any): void {
         this.estacion = estacion;
-        this.dbService.deleteEstacion(this.estacion).subscribe((data: any) => {
-            this.tService.success("Estacion eliminada con exito.", "Envio exitoso");
-            window.location.reload();
-        },
-            (err: any) => {
-                console.log(err);
-                this.tService.error("", "Ha ocurrido un error");
-            });
+        if (confirm("¿Está seguro de eliminar esta estación?")) {
+            this.dbService.deleteEstacion(this.estacion).subscribe((data: any) => {
+                this.tService.success("Estacion eliminada con exito.", "Envio exitoso");
+                window.location.reload();
+            },
+                (err: any) => {
+                    console.log(err);
+                    this.tService.error("", "Ha ocurrido un error");
+                });
+        }
     }
 
 
@@ -197,22 +238,25 @@ export class EstacionesComponent implements OnInit, OnDestroy {
      * @param formEstacion formulario de estación
      */
     submit(formEstacion: NgForm): void {
-        this.dbService.updateEstacion(this.estacion)
-            .subscribe(
-                (data: any) => {
-                    this.tService.success("Estacion actualizada con exito.", "Envio exitoso");
-                    formEstacion.reset();
-                    const table = (<HTMLInputElement>document.getElementById("table"));
-                    const form = (<HTMLInputElement>document.getElementById("form-estacion"));
-                    table.style.display = "block";
-                    form.style.display = "none";
-                    window.location.reload();
-                },
-                (err: any) => {
-                    console.log(err);
-                    this.tService.error("", "Ha ocurrido un error");
-                }
-            );
+        if (confirm("¿Está seguro de actualizar la información de esta estación?")) {
+
+            this.dbService.updateEstacion(this.estacion)
+                .subscribe(
+                    (data: any) => {
+                        this.tService.success("Estacion actualizada con exito.", "Envio exitoso");
+                        formEstacion.reset();
+                        const table = (<HTMLInputElement>document.getElementById("table"));
+                        const form = (<HTMLInputElement>document.getElementById("form-estacion"));
+                        table.style.display = "block";
+                        form.style.display = "none";
+                        window.location.reload();
+                    },
+                    (err: any) => {
+                        console.log(err);
+                        this.tService.error("", "Ha ocurrido un error");
+                    }
+                );
+        }
     }
 
     /**
@@ -227,8 +271,6 @@ export class EstacionesComponent implements OnInit, OnDestroy {
         this.estacion = new Estacion();
         formEstacion.reset();
     }
-
-
 
     editarFoto(estacion: Estacion): void {
         this.estacion = estacion;
@@ -250,7 +292,6 @@ export class EstacionesComponent implements OnInit, OnDestroy {
             );
     }
 
-
     cancelarFoto(): void {
         const table = (<HTMLInputElement>document.getElementById("table"));
         const form = (<HTMLInputElement>document.getElementById("form-update-foto"));
@@ -262,7 +303,45 @@ export class EstacionesComponent implements OnInit, OnDestroy {
 
     subirFoto() {
         if (this.uploader.queue.length != 0) {
-            this.uploader.queue[0].upload()
+            if (confirm("¿Está seguro de actualizar la foto de esta estación?")) {
+                this.uploader.queue[0].upload()
+            }
+        } else {
+            this.tService.error("", "Debe elegir un archivo para actualizar la foto.");
+        }
+    }
+
+    getDivisiones(pais: any) {
+        this.division1 = "";
+        this.division2 = "";
+        this.division3 = "";
+        this.divisiones1 = [];
+        this.divisiones2 = [];
+        this.divisiones3 = [];
+        console.log(pais)
+        this.dbService.getDivisionesPaises(pais)
+            .subscribe((data: any) => {
+                this.divisiones1 = (data as any);
+            });
+    }
+
+    getDivDivisiones(division: any, nivel: number) {
+        if (nivel == 1) {
+            this.division2 = "";
+            this.division3 = "";
+            this.divisiones2 = [];
+            this.divisiones3 = [];
+            this.dbService.getDivDivisiones(division, nivel)
+                .subscribe((data: any) => {
+                    this.divisiones2 = (data as any);
+                });
+        } else if (nivel == 2) {
+            this.division3 = "";
+            this.divisiones3 = [];
+            this.dbService.getDivDivisiones(division, nivel)
+                .subscribe((data: any) => {
+                    this.divisiones3 = (data as any);
+                });
         }
     }
